// src/app/models/masters-mpc/productassociation.ts
export class Productassociation {
    id: number = -1;
    name: string = "";
    createdByUser: string = "";
    createdByUserId: number = -1;
    updatedByUser: string = "";
    updatedByUserId: number = -1;
    createdDate: Date;
    updatedDate: Date;
    active: boolean = false;
    indInitialConfiguration: boolean = false;
}